PDF-443 updated the calculateCIDWidths() method to avoid a nasty class cast exception.

+++ core/src/org/icepdf/core/pobjects/fonts/ofont/Font.java
@@ -711,7 +711,7 @@ public class Font extends org.icepdf.core.pobjects.fonts.Font {
                     for (int j = (Integer) current; j <= (Integer) peek; j++) {
                         // apply Unicode mapping if any
                         currentChar = j;
-                        cidWidths.put(currentChar, (Integer) cidWidth.get(i + 2) / 1000f);
+                        cidWidths.put(currentChar, (Float) cidWidth.get(i + 2) / 1000f);
                     }
                     i += 2;
                 }
